+++ packages/@headlessui-react/CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet!
+### Fixed
+
+- Improve `Menu` component performance ([#3685](https://github.com/tailwindlabs/headlessui/pull/3685))
 
 ## [2.2.1] - 2025-04-04